+++ app/database/__init__.py
@@ -1,0 +1,2 @@
+from .db import assemble_database_url, get_session
+from .models import Restaurant
