+++ app/database/db.py
@@ -1,10 +1,10 @@
 from pydantic import PostgresDsn
 from sqlalchemy.ext.asyncio.engine import AsyncEngine
 from sqlalchemy.orm import sessionmaker
-from sqlmodel import SQLModel, create_engine
+from sqlmodel import create_engine
 from sqlmodel.ext.asyncio.session import AsyncSession
 
-from .config import DatabaseSettings
+from ..config import DatabaseSettings
 
 database_settings = DatabaseSettings()
 
@@ -28,12 +28,6 @@ def assemble_database_url(settings: DatabaseSettings, as_async: bool = True):
 
 
 engine = AsyncEngine(create_engine(assemble_database_url(database_settings), echo=True, future=True))
-
-
-async def init_db():
-    async with engine.begin() as conn:
-        # await conn.run_sync(SQLModel.metadata.drop_all)
-        await conn.run_sync(SQLModel.metadata.create_all)
 
 
 async def get_session():
